Finished Merge and moved QuizSprocket to the proper Extensions folder in the solution.
