+++ Jabbot.sln
@@ -180,10 +180,14 @@ Global
 		{3E8A88F3-737F-4F93-92EA-151B1FE3A44B}.Release|Mixed Platforms.Build.0 = Release|Any CPU
 		{3E8A88F3-737F-4F93-92EA-151B1FE3A44B}.Release|x86.ActiveCfg = Release|Any CPU
 		{93DB29BE-BC75-4D59-9313-891A1245940D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{93DB29BE-BC75-4D59-9313-891A1245940D}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{93DB29BE-BC75-4D59-9313-891A1245940D}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
+		{93DB29BE-BC75-4D59-9313-891A1245940D}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
 		{93DB29BE-BC75-4D59-9313-891A1245940D}.Debug|x86.ActiveCfg = Debug|Any CPU
 		{93DB29BE-BC75-4D59-9313-891A1245940D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{93DB29BE-BC75-4D59-9313-891A1245940D}.Release|Any CPU.Build.0 = Release|Any CPU
 		{93DB29BE-BC75-4D59-9313-891A1245940D}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
+		{93DB29BE-BC75-4D59-9313-891A1245940D}.Release|Mixed Platforms.Build.0 = Release|Any CPU
 		{93DB29BE-BC75-4D59-9313-891A1245940D}.Release|x86.ActiveCfg = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
@@ -197,6 +201,7 @@ Global
 		{B1E7F2FE-7BC0-409A-9BE5-0A0E03553872} = {7170D00A-149E-427F-A7F9-FBA91CE3F9D7}
 		{ED3BAD8A-2972-4D52-BE1E-9B73EE43E53C} = {7170D00A-149E-427F-A7F9-FBA91CE3F9D7}
 		{D05250A1-6F48-49F1-967C-BCB25213EA06} = {7170D00A-149E-427F-A7F9-FBA91CE3F9D7}
+		{93DB29BE-BC75-4D59-9313-891A1245940D} = {7170D00A-149E-427F-A7F9-FBA91CE3F9D7}
 		{3E8A88F3-737F-4F93-92EA-151B1FE3A44B} = {A3A6EC44-1F10-4427-98DC-CF13D11F8C79}
 	EndGlobalSection
 EndGlobal
